# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ChapterResponse:
    properties:
      course_id:
        example: 1
        type: integer
      created_at:
        type: string
      description:
        example: Basic types, declaration, scope.
        type: string
      id:
        example: 1
        type: integer
      lessons:
        items:
          $ref: '#/definitions/dto.LessonResponse'
        type: array
      name:
        example: Variables and Types
        type: string
      order:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
  dto.CourseResponse:
    properties:
      chapters:
        items:
          $ref: '#/definitions/dto.ChapterResponse'
        type: array
      created_at:
        type: string
      description:
        example: Introductory course on Go programming.
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Golang Developer
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateCourseRequest:
    properties:
      description:
        example: Introductory course on Go programming.
        maxLength: 10000
        type: string
      name:
        example: Golang Developer
        maxLength: 255
        minLength: 2
        type: string
    required:
    - name
    type: object
  dto.CreateLessonRequest:
    properties:
      chapter_id:
        example: 1
        type: integer
      content:
        example: In Go, var declares one or more variables...
        type: string
      description:
        example: How to declare variables and infer types.
        maxLength: 10000
        type: string
      name:
        example: Declaring variables with var
        maxLength: 255
        minLength: 2
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
    required:
    - chapter_id
    - name
    type: object
  dto.ErrorBody:
    properties:
      code:
        example: COURSE_NOT_FOUND
        type: string
      message:
        example: course not found
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
    type: object
  dto.LessonResponse:
    properties:
      chapter_id:
        example: 1
        type: integer
      content:
        example: In Go, var declares one or more variables...
        type: string
      created_at:
        type: string
      description:
        example: How to declare variables and infer types.
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Declaring variables with var
        type: string
      order:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateCourseRequest:
    properties:
      description:
        example: Refreshed course on Go programming.
        maxLength: 10000
        type: string
      name:
        example: Golang Developer (v2)
        maxLength: 255
        minLength: 2
        type: string
    required:
    - name
    type: object
  dto.UpdateLessonRequest:
    properties:
      content:
        example: Updated lesson body...
        type: string
      description:
        example: Updated explanation.
        maxLength: 10000
        type: string
      name:
        example: Declaring variables with var (revised)
        maxLength: 255
        minLength: 2
        type: string
      order:
        example: 1
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API for managing Courses, Chapters and Lessons in the LMS platform.
    All mutating endpoints (POST/PUT/DELETE) emit structured logs via logrus.
    Errors are returned in a unified shape: {"error":{"code":"...","message":"..."}}.
  title: LMS Main Service API
  version: "1.0"
paths:
...
  /chapters/{id}/lessons:
    get:
      description: Returns all lessons of the given chapter, ordered by lesson.order.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.LessonResponse'
            type: array
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Chapter not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List lessons of a chapter
      tags:
      - lessons
  /courses:
    get:
      description: Returns a flat list of courses without nested chapters.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CourseResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List all courses
      tags:
      - courses
    post:
      consumes:
      - application/json
      description: Creates a course. Name length 2..255. Description up to 10000 chars.
        Requires ROLE_ADMIN.
      parameters:
      - description: Course payload
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCourseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CourseResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new course (admin only)
      tags:
      - courses
  /courses/{id}:
    delete:
      description: Deletes a course. All chapters and lessons are cascade-deleted.
        Requires ROLE_ADMIN.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a course (admin only)
      tags:
      - courses
    get:
      description: Returns a course together with its chapters, ordered by chapter.order.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseResponse'
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get course by ID (with chapters)
      tags:
      - courses
    put:
      consumes:
      - application/json
      description: Replaces course name and description. Timestamps are preserved.
        Requires ROLE_ADMIN.
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Course payload
        in: body
        name: course
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCourseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CourseResponse'
        "400":
          description: Validation failed or invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Course not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing course (admin only)
      tags:
      - courses
...
  /lessons:
    post:
      consumes:
      - application/json
      description: Creates a lesson under an existing chapter. chapter_id is required
        and validated. Requires ROLE_ADMIN.
      parameters:
      - description: Lesson payload
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLessonRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.LessonResponse'
        "400":
          description: Validation failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Parent chapter not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new lesson (admin only)
      tags:
      - lessons
  /lessons/{id}:
    delete:
      description: Deletes a single lesson. Requires ROLE_ADMIN.
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Lesson not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a lesson (admin only)
      tags:
      - lessons
    get:
      description: Returns a single lesson with its full content.
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonResponse'
        "400":
          description: Invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Lesson not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get lesson by ID
      tags:
      - lessons
    put:
      consumes:
      - application/json
      description: Updates lesson name, content and order. chapter_id cannot be changed
        via PUT. Requires ROLE_ADMIN.
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lesson payload
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLessonRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LessonResponse'
        "400":
          description: Validation failed or invalid id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Requires ROLE_ADMIN
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Lesson not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing lesson (admin only)
      tags:
      - lessons
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"